Renaming util class
